Refactor GardensController#create
Prior problem was that if you entered valid garden data but entered
an invalid planting data, the garden would be created in the db
but user would be sent back to the new view with an error message. User
would resubmit and a *second* instance of that garden would be persisted.
Had to come up with some work arounds, based on whether there were details
in the error messages, to make sure this didn't happen. The main complicating
factor is that when the custom writer is called in the garden model, it must
save or else the nested form cannot be persisted.