Attempt to refactor custom writer in Garden model re update function
Editing/updating works, and if a planting number is invalid data,
the invalid data is not saved.  The problem is that the error is
not highlighted to the user.  Tried to play around with but couldn't
get the error to appear.  Note that when new form is created, this
is not a problem.