Render next garden in garden show page via garden.js -- works
This is a good place to come back to if I need it.

The only strange thing is I find that when creating a new garden,
and not filling in square feet, the garden show page for that new
creation sometime says null.  I played around with the activerecord
validation, and couldn't fix it, but then it went away.  Odd