In species.js, for new species form: rough draft error handling:
In new species form that is rendered, add data-disable false so
that submit button is not disabled after first click with error
message. Also, rearranged species_controller#create so that a
failure to save returns json errors. In the middle of trying to get
those errors to appear in a alert.