In species controller, has idea (not working) for dealing with bug on species add page
Basically see the speciesController #create method. But removing this
because going to live with the bug for now that it is trying to fix --
basically the bug is that when the species form on it's own is submitted
(not in the context of the garden show page), then a json page is
rendered.