Get semantic-ui-rails-LESS to work
Lots of errors in trying to get here.

Steps here would load the JS, but not the stylesheets:
https://stackoverflow.com/questions/37283642/how-do-i-set-up-a-rails-application-with-semantic-ui

So I had to abandon that gem for semantic-ui-rails-less.  Then that wouldn't work due to this error:

ExecJS::RubyRacerRuntime is not supported. Please replace therubyracer with mini_racer in your Gemfile or use Node.js as ExecJS runtime.

Neither of these in the gemfile solved the problem.

gem 'mini_racer'
gem 'mini_racer', platforms: :ruby
(idea from here: https://github.com/ai/autoprefixer-rails/issues/137

The only thing that did solve the problem was adding this to boot.rb:

https://stackoverflow.com/questions/9686868/how-do-i-force-execjs-to-use-node-js

Also, this did not work:

https://github.com/Semantic-Org/Semantic-UI-Rails-LESS/issues/12