In middle of trying to get nested planting new form to work...
...As is, the nested form does not work.  Seems this is impossible in Rails.
Would need a separate form. That's possible. Or would have to use
javascript to hijack the submit button and send the info to the server;
I did this in the prior iteration when creating a new species, but that
was on the user show page, so not a nested form.

One thing I want to try is to see if I can use fields_for and
build a new associated planting.  Let's see...